requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
